+++ babel-defines.js
