+++ package.json
@@ -2,7 +2,6 @@
   "name": "@primer/behaviors",
   "version": "1.0.2",
   "description": "Shared behaviors for JavaScript components",
-  "type": "module",
   "main": "lib/index.js",
   "module": "lib-esm/index.js",
   "exports":{
@@ -29,8 +28,8 @@
     "dist/focus-trap.js"
   ],
   "scripts": {
-    "build:cjs": "cross-env BABEL_MODULE=commonjs babel src --out-dir lib --extensions '.js,.ts,.jsx,.tsx'",
-    "build:esm": "babel src --out-dir lib-esm --extensions '.js,.ts,.jsx,.tsx'",
+    "build:cjs": "cross-env BABEL_MODULE=commonjs babel src --out-dir lib --extensions '.js,.ts'",
+    "build:esm": "babel src --out-dir lib-esm --extensions '.js,.ts'",
     "lint": "eslint src/",
     "test": "npm run jest && npm run lint",
     "test:watch": "jest --watch",
@@ -71,14 +70,9 @@
     "@babel/plugin-proposal-nullish-coalescing-operator": "7.16.0",
     "@babel/plugin-proposal-optional-chaining": "7.14.5",
     "@babel/plugin-transform-modules-commonjs": "7.14.5",
-    "@babel/preset-react": "7.14.5",
     "@babel/preset-typescript": "7.15.0",
     "babel-core": "7.0.0-bridge.0",
     "babel-loader": "^8.2.2",
-    "babel-plugin-add-react-displayname": "0.0.5",
-    "babel-plugin-macros": "3.1.0",
-    "babel-plugin-preval": "5.0.0",
-    "babel-plugin-styled-components": "2.0.2",
     "babel-plugin-transform-replace-expressions": "0.2.0",
     "babel-polyfill": "6.26.0",
     "@changesets/changelog-github": "^0.4.2",
