--- tsconfig.build.json
@@ -1,5 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "exclude": ["script/**/*.ts"]
-  }
-
